requirements.txt
invoke==2.2.1
matplotlib==3.10.7
numpy==2.3.3
pandas==2.3.3
PyYAML==6.0.3
torch==2.8.0
typer==0.19.2
